requirements.txt
# --- Core DB driver ---
psycopg[binary]>=3.2.10,<3.3

# --- Web and utilities ---
websockets==11.0.3
python-dotenv==1.0.1
colorlog==6.8.2
ruff==0.5.7
pytest==7.4.4
matplotlib==3.9.2

# --- Windows optional ---
pywin32; platform_system == "Windows"



flask>=3.0.0
psycopg[binary]>=3.2.10,<3.3

pandas
pandas-market-calendars
